SMF: Fix conversation label import for 2.1
In SMF 2.1 the labels because their own database table instead of being a
comma-separated column somewhere. For purposes of importing the conversation
users we will simulate the old comma-separated storage using `GROUP_CONCAT`.

By prefixing all the label IDs with the user ID we don't need to join in the
`pm_label` table to check the user ID. Any non-matching labels (i.e. labels of
other users) will simply be discarded because the ID with the userID prefix
does not exist.